+++ src/shared/components/layout/Panel/PanelListItemMenuActions.tsx
@@ -1,35 +1,59 @@
 import {
-  FaPenToSquare,
-  FaTrash,
-  FaCrosshairs,
   FaCircleInfo,
-  FaPencil,
+  FaCopy,
+  FaCrosshairs,
   FaDownload,
+  FaLink,
+  FaPencil,
+  FaPenToSquare,
+  FaRegEye,
+  FaTrash,
+  FaUserPlus,
 } from "react-icons/fa6";
 import { MenuButton } from "../Menu/MenuButton";
 import { Separator } from "../Separator";
 
 interface PanelListItemMenuActionsProps {
+  onView?: () => void;
   onCenter?: () => void;
   onDownload?: (() => void) | null;
   onEdit?: () => void;
   onNameChange?: (() => void) | null;
+  onDuplicate?: () => void;
+  onShare?: () => void;
+  onCopytoClipboard?: () => void;
   onRemove?: (() => void) | null;
   removeDisabled?: boolean;
   handleEdit: () => void;
 }
 
 export function PanelListItemMenuActions({
+  onView,
   onCenter,
   onDownload,
   onEdit,
   onNameChange,
+  onDuplicate,
+  onShare,
+  onCopytoClipboard,
   onRemove,
   removeDisabled = false,
   handleEdit,
 }: PanelListItemMenuActionsProps) {
   return (
     <>
+      {onView && (
+        <>
+          <MenuButton
+            onClick={onView}
+            icon={<FaRegEye className="mr-2" />}
+            className="w-full"
+          >
+            View
+          </MenuButton>
+          <Separator className="my-1" />
+        </>
+      )}
       {onDownload && (
         <MenuButton
           onClick={onDownload}
@@ -66,9 +90,37 @@ export function PanelListItemMenuActions({
           Center
         </MenuButton>
       )}
+      {onDuplicate && (
+        <MenuButton
+          onClick={onDuplicate}
+          icon={<FaCopy className="mr-2" />}
+          className="w-full"
+        >
+          Duplicate
+        </MenuButton>
+      )}
+      {(onShare || onCopytoClipboard) && <Separator className="my-1" />}
+      {onShare && (
+        <MenuButton
+          onClick={onShare}
+          icon={<FaUserPlus className="mr-2" />}
+          className="w-full"
+        >
+          Share
+        </MenuButton>
+      )}
+      {onCopytoClipboard && (
+        <MenuButton
+          onClick={onCopytoClipboard}
+          icon={<FaLink className="mr-2" />}
+          className="w-full"
+        >
+          Copy Link
+        </MenuButton>
+      )}
       {onRemove && (
         <>
-          <Separator />
+          <Separator className="my-1" />
           <MenuButton
             onClick={onRemove}
             icon={
@@ -78,10 +130,10 @@ export function PanelListItemMenuActions({
                 <FaTrash className="mr-2" />
               )
             }
-            className={`w-full text-danger ${removeDisabled ? "opacity-50 cursor-not-allowed" : ""}`}
+            className={`w-full !text-danger ${removeDisabled ? "opacity-50 cursor-not-allowed" : ""}`}
             disabled={removeDisabled}
           >
-            {removeDisabled ? "Cannot remove" : "Remove"}
+            {removeDisabled ? "Cannot delete" : "Delete"}
           </MenuButton>
         </>
       )}